signed or unsigned fix

+++ clus/src/PRSegment.cxx
@@ -10,25 +10,25 @@ namespace WireCell::Clus::PR {
     }
 
     int Segment::fit_index(int i){
-        if (i < 0 || i >= m_fits.size()) {
+        if (i < 0 || static_cast<size_t>(i) >= m_fits.size()) {
             throw std::out_of_range("Invalid fit index");
         }
         return m_fits[i].index;
     }
     void Segment::fit_index(int i, int idx){
-        if (i < 0 || i >= m_fits.size()) {
+        if (i < 0 || static_cast<size_t>(i) >= m_fits.size()) {
             throw std::out_of_range("Invalid fit index");
         }
         m_fits[i].index = idx;
     }
     bool Segment::fit_flag_skip(int i){
-        if (i < 0 || i >= m_fits.size()) {
+        if (i < 0 || static_cast<size_t>(i) >= m_fits.size()) {
             throw std::out_of_range("Invalid fit index");
         }
         return m_fits[i].flag_fix;
     }
     void Segment::fit_flag_skip(int i, bool flag){
-        if (i < 0 || i >= m_fits.size()) {
+        if (i < 0 || static_cast<size_t>(i) >= m_fits.size()) {
             throw std::out_of_range("Invalid fit index");
         }
         m_fits[i].flag_fix = flag;
@@ -65,4 +65,3 @@ namespace WireCell::Clus::PR {
 
 
 }
-
